requirements.txt
streamlit
sqlalchemy
pydantic
python-dotenv
openai==0.28.0
numpy
